requirements.txt
# Core dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0

# MIDI
python-rtmidi>=1.5.0
mido>=1.3.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Data visualization
# matplotlib>=3.7.0
# pandas>=2.0.0
